Remove sidebar functionality, and 1 more.
Remove the previous option to show the svn view as a new tab
or as a side bar.  Now we always show it as a tab.  Remove all code
that relates to showing it as a side bar.

Obiwan error in selecting the next item in the list.  It was possible
to go one element too far.